+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ## [Unreleased]
 ### Added
 - FF-2413 - JsonSerializer serialize and deserialize rules
+- FF-2590 - Explicit checks for ArgumentExceptions that they have the parameter name passed
 ### Fixed
 ### Changed
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.9.0.19135